+++ src/Parser/Statements.h
@@ -1,6 +1,8 @@
 #ifndef STATEMENTS_H
 #define STATEMENTS_H
 
+#include <cstdint>
+
 class Statement
 {
 public:
@@ -11,13 +13,13 @@ public:
 class PushStatement : public Statement
 {
 public:
-    PushStatement(int number) : _number(number) {}
+    PushStatement(int64_t number) : _number(number) {}
     ~PushStatement() {}
 
-    int GetNumber() { return _number; }
+    int64_t GetNumber() { return _number; }
 
 private:
-    int _number;
+    int64_t _number;
 };
 
 class DuplicateTopStatement : public Statement
@@ -29,11 +31,11 @@ public:
 class CopyStatement : public Statement
 {
 public:
-    CopyStatement(int number) : _number(number) {}
+    CopyStatement(int64_t number) : _number(number) {}
     ~CopyStatement() {}
 
 private:
-    int _number;
+    int64_t _number;
 };
 
 class SwapStatement : public Statement
@@ -51,11 +53,11 @@ public:
 class SlideStatement : public Statement
 {
 public:
-    SlideStatement(int number) : _number(number) {}
+    SlideStatement(int64_t number) : _number(number) {}
     ~SlideStatement() {}
 
 private:
-    int _number;
+    int64_t _number;
 };
 
 #endif
